servidor/jango/Django_tutorial_bootstrap/requirements.txt
Django==3.1.6
mysqlclient==2.0.3
Pillow==8.1.0
django-ckeditor==6.0.0